# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation de l'API du module de billeterie automatique",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/all_commandes": {
            "get": {
                "tags": [
                    "The_commande"
                ],
                "summary": "Afficher la liste des commandes",
                "description": "Renvoyer la liste de tous les commandes",
                "operationId": "2b5f94f7adb0519d68cd3beabcd84bfe",
                "responses": {
                    "200": {
                        "description": "Les commandes ont été trouvé et leurs informations sont retournées",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Aucune marchandise dans la base de données ",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/commande_show/{id}": {
            "get": {
                "tags": [
                    "The_commande"
                ],
                "summary": "Afficher les infos d'une marchandise",
                "description": "Renvoyer les infos d'une marchandise",
                "operationId": "ad52fa709f232ca0089b5160f250d7f5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Identifiant d'une marchandise à voir ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": " trouvé  !",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Aucun  ne correspond à l'identifiant saisi",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/commande_store": {
            "post": {
                "tags": [
                    "The_commande"
                ],
                "summary": "Enregistrer une nouvelle marchandise",
                "operationId": "c9f0599c83a328eeb2240f17bda9a647",
                "parameters": [
                    {
                        "name": "client_id",
                        "in": "query",
                        "description": "Identifiant d'un client.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Identifiant d'un utilisateur.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "marchandise_id",
                        "in": "query",
                        "description": "Identifiant d'une marchandise.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "destinataire_id",
                        "in": "query",
                        "description": "Identifiant d'un destinataire.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": " enregistre avec succes.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Echec d'enregistrement de l'.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "402": {
                        "description": "Un client avec ces informations existe déja",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/commande_destroy/{id}": {
            "delete": {
                "tags": [
                    "The_commande"
                ],
                "summary": "Supprimer une marchandise",
                "description": "Elle se charger de supprimer un  et donc son compte",
                "operationId": "0647993ce53a2d72e6983c9f88649267",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Identifiant du marchandise à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": " supprimé avec succès!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "L'étudiant n'existe pas !",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    }
}